+++ battlecraft-main/src/main/java/br/com/battlebits/battlecraft/warp/Warp.java
@@ -6,6 +6,7 @@ import br.com.battlebits.battlecraft.event.warp.PlayerWarpQuitEvent;
 import br.com.battlebits.battlecraft.translate.BattlecraftTranslateTag;
 import br.com.battlebits.battlecraft.warp.scoreboard.WarpScoreboard;
 import br.com.battlebits.battlecraft.world.WorldMap;
+import br.com.battlebits.commons.bukkit.event.update.UpdateEvent;
 import br.com.battlebits.commons.command.CommandClass;
 import lombok.Getter;
 import org.bukkit.Location;
@@ -23,6 +24,8 @@ import static br.com.battlebits.battlecraft.translate.BattlecraftTranslateTag.va
 
 @Getter
 public abstract class Warp implements Listener, CommandClass {
+
+    private static final int SCOREBOARD_TICK_UPDATE = 7;
 
     @Getter
     protected Set<Kit> kits;
@@ -93,6 +96,21 @@ public abstract class Warp implements Listener, CommandClass {
             players.remove(event.getPlayer());
             if (players.isEmpty())
                 HandlerList.unregisterAll(this);
+        }
+    }
+
+    @EventHandler
+    public void onTick(UpdateEvent event) {
+        if (event.getType() == UpdateEvent.UpdateType.TICK && event.getCurrentTick() % SCOREBOARD_TICK_UPDATE == 0) {
+            getScoreboard().updateTitleText();
+            for (Player player : getPlayers()) {
+                getScoreboard().updateTitle(player);
+            }
+        }
+        if (event.getType() != UpdateEvent.UpdateType.SECOND)
+            return;
+        for (Player player : getPlayers()) {
+            applyTabList(player);
         }
     }
