+++ battlecraft-main/src/main/java/br/com/battlebits/battlecraft/warp/scoreboard/OneVsOneBoard.java
@@ -1,13 +1,9 @@
 package br.com.battlebits.battlecraft.warp.scoreboard;
 
 import br.com.battlebits.battlecraft.Battlecraft;
-import br.com.battlebits.battlecraft.ability.Kit;
-import br.com.battlebits.battlecraft.manager.KitManager;
 import br.com.battlebits.battlecraft.status.ranking.Queue1v1;
 import br.com.battlebits.battlecraft.status.ranking.RankedQueue;
 import br.com.battlebits.battlecraft.status.warpstatus.Status1v1;
-import br.com.battlebits.battlecraft.status.warpstatus.StatusMain;
-import br.com.battlebits.battlecraft.util.NameUtils;
 import br.com.battlebits.battlecraft.warp.Warp1v1;
 import br.com.battlebits.commons.Commons;
 import br.com.battlebits.commons.CommonsConst;
@@ -18,34 +14,18 @@ import br.com.battlebits.commons.translate.Language;
 import org.bukkit.entity.Player;
 
 import static br.com.battlebits.battlecraft.translate.BattlecraftTranslateTag.*;
-import static br.com.battlebits.battlecraft.translate.BattlecraftTranslateTag.BOARD_MAIN_WEBSITE;
 
-public class OneVsOneBoard implements WarpScoreboard {
+public class OneVsOneBoard extends WarpScoreboard {
 
-    private StringScroller scroller;
-    private String title;
     private Warp1v1 warp;
 
     public OneVsOneBoard(Warp1v1 warp) {
+        super(new StringScroller("BattleCraft - " + warp.getName() + " -", 14, 1));
         this.warp = warp;
-        scroller = new StringScroller("BattleCraft - " + warp.getName() + " -", 14, 1);
-        title = scroller.next();
-    }
-
-    public void updateTitleText() {
-        title = scroller.next();
-    }
-
-    public void updateTitle(Player player) {
-        getBattleBoard(player).update(battleScoreboard -> battleScoreboard.setDisplayName(Commons.getLanguage(player.getUniqueId()).tl(BOARD_MAIN_TITLE, title)));
     }
 
     private RankedQueue getStatus(Player player) {
         return ((Status1v1)Battlecraft.getInstance().getStatusManager().get(player.getUniqueId()).getWarpStatus(warp)).getQueueStatus(Queue1v1.NORMAL);
-    }
-
-    private BattleScoreboard getBattleBoard(Player player) {
-        return ((BukkitAccount) Commons.getAccount(player.getUniqueId())).getBattleboard();
     }
 
     public void updateVictory(Player player) {
@@ -68,16 +48,11 @@ public class OneVsOneBoard implements WarpScoreboard {
 
     @Override
     public void applyScoreboard(Player player) {
+        super.applyScoreboard(player);
         BukkitAccount account = (BukkitAccount) Commons.getAccount(player.getUniqueId());
         RankedQueue main = getStatus(player);
         Language l = account.getLanguage();
         BattleScoreboard board = account.getBattleboard();
-        board.getLines().clear();
-        board.update();
-        board.setDisplayName(l.tl(BOARD_MAIN_TITLE, title));
-        String kitName = account.getLanguage().tl(KIT_NONE);
-        if (KitManager.containsKit(player))
-            kitName = NameUtils.formatString(KitManager.getCurrentPlayerKit(player).getName());
         int nullLine = 0;
         board.addLine("n" + ++nullLine, "");
         board.addLine("victory", l.tl(BOARD_1v1_VICTORY, main.getVictory()));
